Rework `Avatar` indicator to be more generic
